# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "version": "1.0.0"
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "consumes": [
        "multipart/form-data"
    ],
    "paths": {
...
        "/api/auth/logout": {
            "post": {
                "summary": "New Request",
                "description": "",
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    }
                },
                "tags": [
                    "deconnexion"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Authorization",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "User-Agent",
                        "type": "string"
                    },
                    {
                        "in": "header",
                        "name": "bearer",
                        "type": "string"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "access_token": {
                                        "type": "string"
                                    },
                                    "token_type": {
                                        "type": "string"
                                    },
                                    "expires_in": {
                                        "type": "integer"
                                    }
                                }
                            },
                            "example": {
                                "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDEvYXBpL2F1dGgvbG9naW4iLCJpYXQiOjE3MjMwMDc3NDIsImV4cCI6MTcyMzAwOTQyMiwibmJmIjoxNzIzMDA3NzQyLCJqdGkiOiJMUVp3S2tVTTFueUtyTG1zIiwic3ViIjoiMTIiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.DJzurkGhwvcmmSUdpOblALPZEFyZiiqVRgCXttoKREI",
                                "token_type": "bearer",
                                "expires_in": 1680
                            }
                        }
                    }
                }
            }
        },
...
    }
}